+++ src/resolve/root.h
@@ -9,7 +9,7 @@ int root_resolve(Node* head);
 
 int root_degree(Node* head, int* degree);
 int root_get_degree(Node* head, int* degree);
-int root_coefficients(Node* head, long double* coefficients, int degree);
+int root_coefficients(Node* head, double* coefficients, int degree);
 
 int root_second_degree(Node* head);
 int root_third_degree(Node* head);
