+++ src/token/token.h
@@ -6,14 +6,14 @@
 #include "stdio.h"
 
 typedef struct _Token {
-    long double value;
+    double value;
     int type;
     struct _Token* next;
 } Token;
 
 Token* t_construct();
 Token* t_destruct(Token* head);
-int t_push(Token* head, int type, long double value);
+int t_push(Token* head, int type, double value);
 int t_next(Token** curent);
 void t_print(Token* head);
